Use realistic MSP430-derived energy costs for IR estimator
Change IRBasedEstimator instruction costs from int to double to support
fractional nanojoule values. Update sample_energy_config_ir.json with
costs derived from MSP430 assembly measurements. Replace the single
sample_milp_config.json with four capacitor-specific configs (100nF,
1uF, 10uF, 100uF) at 3.6V-1.8V voltage range.

+++ passes/include/estimator/IRBasedEstimator.h
@@ -24,14 +24,14 @@ private:
     /// Private constructor - use create() factory method.
     IRBasedEstimator() = default;
 
-    std::unordered_map<std::string, int> instructionCosts_;
+    std::unordered_map<std::string, double> instructionCosts_;
 
     /// Load configuration from JSON file.
     /// @return true on success, false on error (with message to errs()).
     bool loadConfig(const std::string &path);
 
     /// Get cost for an LLVM opcode.
-    int getInstructionCost(unsigned Opcode) const;
+    double getInstructionCost(unsigned Opcode) const;
 
     /// Map opcode to cost category name.
     std::string getCostCategory(unsigned Opcode) const;
